chore: rename model file
todo: Command structure revamp! >> clean up command registry
