+++ routes/home/guest/Documents/index.tsx
@@ -1,6 +1,6 @@
 import { Handlers, PageProps } from "$fresh/server.ts";
 import { TerminalPrompt } from "../../../../islands/terminal/Terminal.tsx";
-import { IDirectoryItem } from "../../../../src/models/Command.ts";
+import { IDirectoryItem } from "../../../../src/models/command.model.ts";
 
 export const handler: Handlers = {
   GET(req, ctx) {
