+++ src/models/command.model.ts
